add pages, navbar links
