+++ client/src/pages/FriendsNmessage.tsx
@@ -1,0 +1,14 @@
+import React, { useState, useRef } from "react";
+const FriendsNmessage = () => {
+    return (
+        <>
+            {
+                <div>
+                    <h1>Friends and Message Page</h1>
+                </div>    
+            }
+        </>
+    );
+}
+
+export default FriendsNmessage;
