+++ client/src/pages/Login.tsx
@@ -1,0 +1,13 @@
+import React, { useState, useRef } from "react";
+const Login = () => {
+    return (
+        <>
+            {
+                <div>
+                    <h1>Login Page</h1>
+                </div>    
+            }
+        </>
+    );
+}
+export default Login;
