+++ client/src/pages/Logout.tsx
@@ -1,0 +1,13 @@
+import React, { useState, useRef } from "react";
+const Logout = () => {
+    return (
+        <>
+            {
+                <div>
+                    <h1>Logout Page</h1>
+                </div>    
+            }
+        </>
+    );
+}
+export default Logout;
