+++ client/src/pages/Shopping.tsx
@@ -1,0 +1,13 @@
+import React, { useState, useRef } from "react";
+const Shopping = () => {
+    return (
+        <>
+            {
+                <div>
+                    <h1>Shopping Page</h1>
+                </div>    
+            }
+        </>
+    );
+}
+export default Shopping;
